modified:   main/gui/0-menu.c 	modified:   main/gui/1-console.c 	modified:   main/gui/1-hex.c 	modified:   main/gui/2-logic.c 	modified:   main/gui/2-real.c 	modified:   main/gui/3-d.c 	modified:   main/gui/3-jiong.c 	modified:   main/gui/gui.c 	modified:   wrap/win/window.c

+++ main/gui/3-d.c
@@ -19,7 +19,7 @@ tempprint()
 		}
 	}
 }
-tempmessage(DWORD type,DWORD key)
+tempmessage(QWORD type,QWORD key)
 {
 	
 }
